Change back to long long in fileio_internal.hpp, and included -Wno-long-long in Makeheader.
git-svn-id: https://code.nr.no/scm/svn/crava/trunk@33 176de4d3-4747-0410-aa96-47f6d6634214

+++ lib/fileio_internal.hpp
@@ -16,11 +16,7 @@
 static const unsigned int masks[4] = 
 {0x000000ff, 0x0000ff00, 0x00ff0000, 0xff000000};
 
-//
-// NBNB-PAL: Does not compile on Linux
-//
-//static const unsigned long long lmasks[8] = 
-static const unsigned long lmasks[8] = 
+static const unsigned long long lmasks[8] = 
 {0x00000000000000ff, 0x000000000000ff00, 
  0x0000000000ff0000, 0x00000000ff000000,
  0x000000ff00000000, 0x0000ff0000000000,
@@ -33,11 +29,7 @@ namespace NRLib2_internal {
   };
 
   union DoubleAsLongLong {
-    //
-    // NBNB-PAL: Does not compile on Linux
-    //
-    //    /* uint64_t */ unsigned long long ull;
-    /* uint64_t */ unsigned long ull;
+    /* uint64_t */ unsigned long long ull;
     double d;
   };
